requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
cupy-cuda11x>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
pillow>=9.0.0
ipython>=8.0.0  # For notebook display